Initial commit: server-config plugin
This plugin enables access (download and upload) to the server config
files.  It may be used to change Gerrit config files (like
etc/gerrit.config) in cases where direct access to the file system where
Gerrit's config files are stored is difficult or impossible to get.

This plugin requires the administrateServer capability for granting
the access to the server config files.

It is recommended to enable this plugin only when it is needed and,
otherwise, keep this plugin disabled. This means that an admin would
typically do the following command sequence:

  $ ssh admin@server gerrit plugin enable server-config
  $ curl --digest -n http://server/a/plugins/server-config/etc/gerrit.config >gerrit.config
  ... edit the gerrit.config ...
  $ curl -n --digest -X PUT --data-binary @gerrit.config http://server/a/plugins/server-config/etc/gerrit.config
  $ ssh admin@server gerrit plugin remove server-config

Change-Id: Ib806cfc6a8ca8645c2d8616c7e963f4c1f5f9722