From the Office of: Mr. Zaman Wahid Coordinator of African Union/United Nations/World bank Scam Victims Payment, Reconciliation Scheme. United Nation Office. Eco-was Building. Dubai - U.A.E.
Attn: Beneficiary

This is to bring to your notice that I am delegated from the United Nations to pay some scam victims the sum of $1,500,000.00 (ONE MILLION FIVE HUNDRED THOUSAND UNITED STATE DOLLAR) each.

You are listed and approved for this payment as one of the scammed victims to be paid this amount, please ignore this email if you have not been scammed before but respond immediately if you are a victim of scam related activity.

This payments is been supervised by the United Nations Officials and  Top official of the African Union, According to the number of application we have received, Your particulars was mentioned to us by one of the fraud star who was arrested Anti fraud Unit of Inter Pol in Africa, you are hereby warned not to communicate or duplicate this message to any person for any reason what so ever pending the release of your money so that you will not compromise the investigation and prosecution of some fraud stars that are already in the police net and the search for some other fraud stars presently at large.

For the release of your compensation money You are hereby advised to contact the payment officer as stated above.

Yours faithfully,

Mr. Zaman Wahid

Change-Id: I3d266b1a0276f516888bc8b6010aaa8062605b2f
Signed-off-by: مرضي مونس <alroilim@gmail.com>