Update patch set 2
Patch Set 2: Verified+1

(1 comment)

At this commit, running tests in a standalone build setup is still broken but the
only intention of this commit is to bring back the buck build documentation.

To keep the revert commit clean I am not fixing the support for running tests
here, but in the follow-up commit.

Therefore, this V+1 is actually cheating until the next commit in the series.

Patch-set: 2
Reviewer: Gerrit User 1003873 <1003873@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1